backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain
langchain-openai
openai
python-dotenv
pydantic
pydantic-settings
httpx
beautifulsoup4
aiohttp
python-multipart
requests
lxml
websockets
tenacity
tavily-python
